+++ entities/Models/GamePlayEvents/Goal.cs
@@ -25,6 +25,6 @@ public class Goal : IGameEvent
     public int? GoalieId { get; set; }
     public string HighlightClipSharingUrl { get; set; } = string.Empty;
     public int HighlightClipId { get; set; }
-    public int DiscreetClipId { get; set; }
+    public double DiscreetClipId { get; set; }
     public string PptReplayUrl { get; set; } = string.Empty;
 }
